backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
google-generativeai>=0.3.2
langchain>=0.2.0
langchain-google-genai>=0.0.10
langchain-huggingface>=0.0.10
langchain-community>=0.0.27
langchain-text-splitters>=0.0.1
chromadb>=0.5.0
sentence-transformers>=2.6.0
torch>=2.3.0
# scikit-learn
python-multipart
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
tabulate>=0.9.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
requests>=2.31.0
aiohttp>=3.9.1
websockets>=12.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
pytest-httpx>=0.29.0
bs4>=0.0.2